backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
gunicorn==23.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
sqlalchemy-cockroachdb==2.0.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
PyJWT==2.9.0
google-auth==2.35.0
boto3==1.35.36
requests==2.28.1
pandas==2.2.3
numpy==2.1.1
bcrypt==4.2.0
scikit-learn==1.5.2
razorpay==1.3.0
